% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_eqret.R
\name{bl_eqret}
\alias{bl_eqret}
\title{Black-Litterman: Retorno de equilibrio}
\usage{
bl_eqret(ra, cov, mktcapw, rf)
}
\arguments{
\item{ra}{Aversion al riesgo}

\item{cov}{Matriz de varianza y covarianza}

\item{mktcapw}{Capitalización de mercado}

\item{rf}{Tasa libre de riesgo}
}
\value{

}
\description{
Black-Litterman: Retorno de equilibrio
}
