% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_optimal.R
\name{bl_optimal}
\alias{bl_optimal}
\title{Black-Litterman: Optimal Weights}
\usage{
bl_optimal(ra, mu, cov)
}
\arguments{
\item{ra}{Aversión al riesgo}

\item{mu}{Retornos}

\item{cov}{Matriz varianza y covarianza}
}
\value{

}
\description{
Black-Litterman: Optimal Weights
}
