% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_portfolio.R
\name{bl_portfolio}
\alias{bl_portfolio}
\title{Black Litterman: Portfolio}
\usage{
bl_portfolio(bl_er, bl_w, bl_cov)
}
\arguments{
\item{bl_er}{Retorno esperado por BL}

\item{bl_w}{Pesos por BL}

\item{bl_cov}{Matriz varianza covarianza por BL}
}
\value{

}
\description{
Black Litterman: Portfolio
}
