% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_posterior.R
\name{bl_posterior}
\alias{bl_posterior}
\title{Black-Litterman: Retorno esperado}
\usage{
bl_posterior(mu, cov, pmat = NULL, qmat, tau, omega = NULL, confidence = NULL)
}
\arguments{
\item{mu}{something}

\item{cov}{something}

\item{pmat}{something}

\item{qmat}{something}

\item{tau}{something}

\item{omega}{something}

\item{confidence}{something}
}
\value{

}
\description{
Black-Litterman: Retorno esperado
}
