% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_risk_aversion.R
\name{bl_risk_aversion}
\alias{bl_risk_aversion}
\title{Black-Litterman: Risk Aversion}
\usage{
bl_risk_aversion(bench, rf)
}
\arguments{
\item{bench}{Activo benchmark}

\item{rf}{Tasa libre de riesgo}
}
\value{

}
\description{
Black-Litterman: Risk Aversion
}
