% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_summary.R
\name{bl_summary}
\alias{bl_summary}
\title{Black-Litterman: Resumen}
\usage{
bl_summary(mkt_cap, new_weights)
}
\arguments{
\item{mkt_cap}{Capitalización de mercado}

\item{new_weights}{Nuevos pesos}
}
\value{

}
\description{
Black-Litterman: Resumen
}
