% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covmat}
\alias{covmat}
\title{Matriz variaza-covarianza.}
\format{An object of class \code{matrix} with 30 rows and 30 columns.}
\usage{
covmat
}
\description{
Matriz variaza-covarianza.
}
\keyword{datasets}
