% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{cumulative_ggplot}
\alias{cumulative_ggplot}
\title{Grafica Portafolios}
\usage{
cumulative_ggplot(returns_tbl, title, subtitle, caption)
}
\arguments{
\item{returns_tbl}{Retornos}

\item{title}{Titulo}

\item{subtitle}{Subtitulo}

\item{caption}{Nota al pie}
}
\value{

}
\description{
Grafica Portafolios
}
