% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Estadistica Descriptiva}
\usage{
descriptive_stats(return_tbl)
}
\arguments{
\item{return_tbl}{Retornos}
}
\value{

}
\description{
Estadistica Descriptiva
}
