% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stocks.R
\name{download_stocks}
\alias{download_stocks}
\title{Descarga acciones}
\usage{
download_stocks(stocks, periodicity, from, to)
}
\arguments{
\item{stocks}{Vectores con el nombre de los tickers}

\item{periodicity}{Frecuencia}

\item{from}{Desde}

\item{to}{Hasta}
}
\value{

}
\description{
Descarga acciones
}
