% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_frontier.R
\name{efficient_frontier}
\alias{efficient_frontier}
\title{Frontera Eficiente}
\usage{
efficient_frontier(
  er,
  cov.mat,
  nport = 20,
  alpha.min = -0.5,
  alpha.max = 1.5,
  shorts = TRUE
)
}
\arguments{
\item{er}{Retornos esperados}

\item{cov.mat}{Matriz de varianza-covarianza}

\item{nport}{nport, el valor por defecto es 20}

\item{alpha.min}{alpha min, el valor por defecto es -0.5}

\item{alpha.max}{alpha max el valor por defecto es 1.5}

\item{shorts}{Si es TRUE incluye venta corta}
}
\value{

}
\description{
Frontera Eficiente
}
