% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_portfolio.R
\name{efficient_portfolio}
\alias{efficient_portfolio}
\title{Portafolio Eficiente}
\usage{
efficient_portfolio(er, cov.mat, target.return, shorts = TRUE)
}
\arguments{
\item{er}{Retornos}

\item{cov.mat}{Matriz de varianza-covarianza}

\item{target.return}{Retorno objetivo}

\item{shorts}{Si es TRUE incluye venta corta}
}
\value{

}
\description{
Portafolio Eficiente
}
