% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPorfolio.R
\name{getPortfolio}
\alias{getPortfolio}
\title{Portafolio}
\usage{
getPortfolio(er, cov.mat, weights)
}
\arguments{
\item{er}{Retornos esperados}

\item{cov.mat}{Matriz de varianza-covarianza}

\item{weights}{Ponderaciones}
}
\value{

}
\description{
Portafolio
}
