% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalMin_portfolio.R
\name{globalMin_portfolio}
\alias{globalMin_portfolio}
\title{Portafolio Global Mínimo}
\usage{
globalMin_portfolio(er, cov.mat, shorts = TRUE)
}
\arguments{
\item{er}{Retorno esperado}

\item{cov.mat}{Matriz de varianza-covarianza}

\item{shorts}{Si es TRUE incluye venta corta}
}
\value{

}
\description{
Portafolio Global Mínimo
}
