% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cov_mat.R
\name{import_cov_mat}
\alias{import_cov_mat}
\title{Importar matriz varianza-covarianza}
\usage{
import_cov_mat(cov_mat)
}
\arguments{
\item{cov_mat}{something}
}
\value{

}
\description{
Importar matriz varianza-covarianza
}
