% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_investor_views.R
\name{import_investor_views}
\alias{import_investor_views}
\title{Importar visiones de los inversionistas}
\usage{
import_investor_views(investor_views)
}
\arguments{
\item{investor_views}{something}
}
\value{

}
\description{
Importar visiones de los inversionistas
}
