% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mkt_cap.R
\name{import_mkt_cap}
\alias{import_mkt_cap}
\title{Importa Market Capitalization}
\usage{
import_mkt_cap(mkt_cap)
}
\arguments{
\item{mkt_cap}{something}
}
\value{

}
\description{
Importa Market Capitalization
}
