% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mktcap}
\alias{mktcap}
\title{Capitalizacion de mercado.}
\format{An object of class \code{matrix} with 1 rows and 30 columns.}
\usage{
mktcap
}
\description{
Capitalizacion de mercado.
}
\keyword{datasets}
