% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_return.R
\name{monthly_return}
\alias{monthly_return}
\title{Retorno Mensual}
\usage{
monthly_return(stocks_tbl, type = c("log", "arithmetic"))
}
\arguments{
\item{stocks_tbl}{Activos}

\item{type}{Tipo de retorno, log = "logaritmo" y arithmetic = aritmetico}
}
\value{

}
\description{
Retorno Mensual
}
