% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{portfolio_all_ggplot}
\alias{portfolio_all_ggplot}
\title{Grafica Portafolios Markowitz & Black-Litterman}
\usage{
portfolio_all_ggplot(log_returns_tbl, bl_port, rf, title, subtitle, caption)
}
\arguments{
\item{log_returns_tbl}{Retornos ln}

\item{bl_port}{Portafolio BL}

\item{rf}{Tasa libre de riesgo}

\item{title}{Titulo}

\item{subtitle}{Subtitulo}

\item{caption}{Nota al pie}
}
\value{

}
\description{
Grafica Portafolios Markowitz & Black-Litterman
}
