% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{portfolio_bl_ggplot}
\alias{portfolio_bl_ggplot}
\title{Grafica Portafolios Black-Litterman}
\usage{
portfolio_bl_ggplot(
  stats_port,
  min_port,
  eff_port,
  eff_front,
  tang_port,
  bl_port,
  rf,
  title,
  subtitle,
  caption
)
}
\arguments{
\item{stats_port}{Promedio, desviacion estandar y varianza-covarianza}

\item{min_port}{Portafolio minima varianza}

\item{eff_port}{Portafolio eficiente sujeto a un retorno}

\item{eff_front}{Frontera efciente}

\item{tang_port}{Portafolio tangente}

\item{bl_port}{Portafolio BL}

\item{rf}{Tasa libre de riesgo}

\item{title}{Titulo}

\item{subtitle}{Subtitulo}

\item{caption}{Nota al pie}
}
\value{

}
\description{
Grafica Portafolios Black-Litterman
}
