% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary_portfolio.R
\name{preliminary_portfolio}
\alias{preliminary_portfolio}
\title{Preliminares para el portafolio}
\usage{
preliminary_portfolio(returns)
}
\arguments{
\item{returns}{Dataframe o tibble con los retornos}
}
\value{

}
\description{
Preliminares para el portafolio
}
