% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangency_portfolio.R
\name{tangency_portfolio}
\alias{tangency_portfolio}
\title{Portfolio Tangente}
\usage{
tangency_portfolio(er, cov.mat, risk.free, shorts = TRUE)
}
\arguments{
\item{er}{Retorno esperado}

\item{cov.mat}{Matriz de varianza-covarianza}

\item{risk.free}{Tasa libre de riesgo}

\item{shorts}{Si es TRUE incluye venta corta}
}
\value{

}
\description{
Portfolio Tangente
}
