% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_posterior.R
\name{bl_posterior}
\alias{bl_posterior}
\title{Black-Litterman: Retorno esperado}
\usage{
bl_posterior(mu, cov, pmat = NULL, qmat, tau, omega = NULL, confidence = NULL)
}
\arguments{
\item{mu}{Retorno}

\item{cov}{Matrix varianza y covarianza}

\item{pmat}{Matriz P}

\item{qmat}{Matriz Q}

\item{tau}{Escalar tau}

\item{omega}{Matriz Omega}

\item{confidence}{Nivel de confianza}
}
\value{

}
\description{
Black-Litterman: Retorno esperado
}
